move pagination to kernel
