+++ application/rest-api-adapter/src/main/java/onlydust/com/marketplace/api/rest/api/adapter/mapper/ProjectRewardMapper.java
@@ -3,7 +3,7 @@ package onlydust.com.marketplace.api.rest.api.adapter.mapper;
 import onlydust.com.marketplace.api.contract.model.*;
 import onlydust.com.marketplace.api.domain.view.ProjectRewardView;
 import onlydust.com.marketplace.api.domain.view.ProjectRewardsPageView;
-import onlydust.com.marketplace.api.domain.view.pagination.PaginationHelper;
+import onlydust.com.marketplace.kernel.pagination.PaginationHelper;
 
 import java.util.Objects;
