+++ domain/src/main/java/onlydust/com/marketplace/api/domain/port/input/BackofficeFacadePort.java
@@ -2,7 +2,7 @@ package onlydust.com.marketplace.api.domain.port.input;
 
 import onlydust.com.marketplace.api.domain.model.Ecosystem;
 import onlydust.com.marketplace.api.domain.view.backoffice.*;
-import onlydust.com.marketplace.api.domain.view.pagination.Page;
+import onlydust.com.marketplace.kernel.pagination.Page;
 
 import java.util.List;
 import java.util.UUID;
