+++ domain/src/main/java/onlydust/com/marketplace/api/domain/port/input/ProjectFacadePort.java
@@ -2,8 +2,8 @@ package onlydust.com.marketplace.api.domain.port.input;
 
 import onlydust.com.marketplace.api.domain.model.*;
 import onlydust.com.marketplace.api.domain.view.*;
-import onlydust.com.marketplace.api.domain.view.pagination.Page;
-import onlydust.com.marketplace.api.domain.view.pagination.SortDirection;
+import onlydust.com.marketplace.kernel.pagination.Page;
+import onlydust.com.marketplace.kernel.pagination.SortDirection;
 import org.apache.commons.lang3.tuple.Pair;
 
 import java.io.InputStream;
