--- domain/src/main/java/onlydust/com/marketplace/api/domain/view/pagination/SortDirection.java
@@ -1,5 +1,0 @@
-package onlydust.com.marketplace.api.domain.view.pagination;
-
-public enum SortDirection {
-    asc, desc
-}
