+++ kernel/src/main/java/onlydust/com/marketplace/kernel/pagination/SortDirection.java
@@ -1,0 +1,5 @@
+package onlydust.com.marketplace.kernel.pagination;
+
+public enum SortDirection {
+    asc, desc
+}
